kendra-bedrock-sharepoint-cdk-python/requirements.txt
aws-cdk-lib==2.110.0
constructs>=10.0.0,<11.0.0
boto3>=1.28.0
cdk-nag>=2.27.0
cfn-lint>=0.83.0